Market-ready hardening: security, payments, production infrastructure
Phase 1 — Security:
- Fix XSS in HTMX f-string HTML → Jinja2 template partials
- Add CSRF middleware (itsdangerous-signed tokens) on all form POSTs
- Enforce secret key validation in production (debug=False)
- Whitelist sort_by columns to prevent injection via getattr
- Add rate limiting middleware on auth endpoints (5 req/min per IP)

Phase 2 — Data Integrity & Flows:
- Float → Numeric(10,2) for all monetary columns (orders + listings)
- Add profile edit, order detail, listing edit, error pages (404/500)
- CSRF tokens in all form templates, profile link in nav
- Decimal-based fee calculation in order service

Phase 3 — Stripe Connect:
- Full onboarding flow (create account → AccountLink → callback)
- Webhook handler (payment_intent.succeeded/failed, account.updated)
- PaymentIntent creation on order, Transfer on completion
- Stripe setup section on profile page

Phase 4 — Production:
- Structured JSON logging (debug: readable, prod: JSON)
- Production config validation (secret key, DB warnings)
- PostgreSQL connection pooling (pool_size, pre_ping)
- Docker + docker-compose (app + PostgreSQL)
- Enhanced health check (DB connectivity, Stripe status)
- 42 tests passing (was 29): +security, +order lifecycle, +CSRF tests

+++ app/models/order.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import String, Float, DateTime, Text, ForeignKey, JSON
+from sqlalchemy import String, Float, DateTime, Text, ForeignKey, JSON, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -25,10 +25,10 @@ class Order(Base):
 
     # Order details
     quantity: Mapped[float] = mapped_column(Float, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Float, nullable=False)
-    total_price: Mapped[float] = mapped_column(Float, nullable=False)
-    platform_fee: Mapped[float] = mapped_column(Float, default=0.0)
-    seller_payout: Mapped[float] = mapped_column(Float, default=0.0)
+    unit_price: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False)
+    total_price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    platform_fee: Mapped[float] = mapped_column(Numeric(10, 2), default=0.0)
+    seller_payout: Mapped[float] = mapped_column(Numeric(10, 2), default=0.0)
     currency: Mapped[str] = mapped_column(String(3), default="USD")
 
     # Status: pending -> accepted -> active -> completed | cancelled | disputed
